# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API  de OZ STREAM",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8001/",
            "description": "OZ STREAM "
        }
    ],
    "paths": {
        "/api/upload/media": {
            "post": {
                "tags": [
                    "Upload Media"
                ],
                "summary": "Upload the media",
                "description": "return an url",
                "operationId": "uploadMedia",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "media"
                                ],
                                "properties": {
                                    "media_id": {
                                        "type": "string"
                                    },
                                    "saison_id": {
                                        "type": "string"
                                    },
                                    "isFilmBande": {
                                        "type": "string"
                                    },
                                    "media": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/getMediaBandeData/{media_id}": {
            "post": {
                "tags": [
                    "media streamming"
                ],
                "summary": "Uget datas streams",
                "description": "return the content",
                "operationId": "streamming",
                "parameters": [
                    {
                        "name": "media_id",
                        "in": "path",
                        "description": "id of media",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "20110": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/getMediaData/{media_id}": {
            "post": {
                "tags": [
                    "media streamming"
                ],
                "summary": "Uget datas streams",
                "description": "return the content",
                "operationId": "streamming",
                "parameters": [
                    {
                        "name": "media_id",
                        "in": "path",
                        "description": "id of media",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "access",
                        "in": "path",
                        "description": "user token on base64",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "20110": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/getSaisonBandeData/{saison_id}": {
            "post": {
                "tags": [
                    "media streamming"
                ],
                "summary": "Uget datas streams",
                "description": "return the content",
                "operationId": "streamming",
                "parameters": [
                    {
                        "name": "saison_id",
                        "in": "path",
                        "description": "id of saison",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "20110": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}